requirements.txt
Pillow==4.1.1
terminaltables==3.1.0